Backend/requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
psycopg2-binary==2.9.1
sqlalchemy==1.4.23
pydantic==1.8.2
bcrypt==4.0.1
paramiko==3.3.1
